pyproject.toml
[tool.poetry]
name = "gpt-jobhunter"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "jobhunter"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
beautifulsoup4 = "4.11.2"
nltk = "3.8.1"
coverage = "5.5"
mypy = "1.6.1"
pandas = "1.5.3"
python-dotenv = "1.0.0"
scikit-learn = "^1.3.0"
streamlit = "^1.52.0"
tqdm = "4.65.0"
gensim = "4.3.2"
PyPDF2 = "1.27.0"
openai = "^1.70.0"
scipy = "1.11.2"
pdfplumber = "^0.11.6"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
pydantic = "^2.5.0"
python-multipart = "^0.0.6"
requests = "^2.31.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
email-validator = "^2.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"

[tool.poetry.scripts]
jobhunter = "jobhunter.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
